backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
plotly==5.18.0
openpyxl==3.1.2
python-dotenv==1.0.0
pydantic==2.5.3 